cross-platform use of int and long is condemnable. segfault fixed
git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@699 1dbd42e5-0d20-4692-8756-183d5fc073b5

+++ modules/c++/sys/include/sys/Conf.h
@@ -204,7 +204,7 @@ namespace sys
             return;
 
         unsigned short half = elemSize >> 1;
-        long offset = 0, innerOff = 0, innerSwap = 0;
+        size_t offset = 0, innerOff = 0, innerSwap = 0;
 
         for(size_t i = 0; i < numElems; ++i, offset += elemSize)
         {
